Crate spinning

Structs

Lazy

A value which is initialized on the first access.

Once

A synchronization primitive which initializes a value lazily, once. Since this also includes a value, it is a bit more like Once from parking_lot or spin.

RawMutex

An extremely simple spinlock, locking by compare-and-swapping a single flag and repeating.

RawRwLock

Enums

OnceState

Type Definitions

MappedMutexGuard

MappedRwLockReadGuard

MappedRwLockWriteGuard

Mutex

MutexGuard

ReentrantMutex

ReentrantMutexGuard

RwLock

RwLockReadGuard

RwLockUpgradableReadGuard

RwLockWriteGuard